+++ src/accessory/VideoLockAccessory.ts
@@ -101,7 +101,6 @@ export default class VideoLockAccessory extends BaseAccessory {
     const doorbellStatus = status.find(_status => _status.code === doorbellSchema.code);
     if (doorbellStatus && doorbellStatus.value === true && this.intialized) {
       this.log.info('Doorbell ring detected.');
-      this.stream?.prewarmSnapshot();
       this.getDoorbellService().getCharacteristic(this.Characteristic.ProgrammableSwitchEvent)
         .updateValue(0); // SINGLE_PRESS
     }